requirements/dev.txt
-r base.txt

# Testing
model_bakery==1.12.0
pytest-django==4.5.2
pytest-cov==4.1.0

# Linter
autoflake==1.4
black==22.3.0
isort==5.10.1
flake8==4.0.1
flake8-bugbear==22.4.25
flake8-comprehensions==3.10.0
flake8-eradicate==1.2.1
flake8-pie==0.15.0
pyflakes==2.4.0

djangorestframework-simplejwt==5.2.2
